+++ hutool-core/src/main/java/cn/hutool/core/thread/DelegatedExecutorService.java
@@ -1,5 +1,7 @@
 package cn.hutool.core.thread;
 
+import cn.hutool.core.lang.Assert;
+
 import java.util.Collection;
 import java.util.List;
 import java.util.concurrent.AbstractExecutorService;
@@ -18,69 +20,79 @@ import java.util.concurrent.TimeoutException;
 public class DelegatedExecutorService extends AbstractExecutorService {
 	private final ExecutorService e;
 
+	/**
+	 * 构造
+	 *
+	 * @param executor {@link ExecutorService}
+	 */
 	DelegatedExecutorService(ExecutorService executor) {
+		Assert.notNull(executor, "executor must be not null !");
 		e = executor;
 	}
 
-	@SuppressWarnings("NullableProblems")
+	@Override
 	public void execute(Runnable command) {
 		e.execute(command);
 	}
 
+	@Override
 	public void shutdown() {
 		e.shutdown();
 	}
 
-	@SuppressWarnings("NullableProblems")
+	@Override
 	public List<Runnable> shutdownNow() {
 		return e.shutdownNow();
 	}
 
+	@Override
 	public boolean isShutdown() {
 		return e.isShutdown();
 	}
 
+	@Override
 	public boolean isTerminated() {
 		return e.isTerminated();
 	}
 
-	@SuppressWarnings("NullableProblems")
+	@Override
 	public boolean awaitTermination(long timeout, TimeUnit unit) throws InterruptedException {
 		return e.awaitTermination(timeout, unit);
 	}
 
-	@SuppressWarnings("NullableProblems")
+	@Override
 	public Future<?> submit(Runnable task) {
 		return e.submit(task);
 	}
 
-	@SuppressWarnings("NullableProblems")
+	@Override
 	public <T> Future<T> submit(Callable<T> task) {
 		return e.submit(task);
 	}
 
-	@SuppressWarnings("NullableProblems")
+	@Override
 	public <T> Future<T> submit(Runnable task, T result) {
 		return e.submit(task, result);
 	}
 
-	@SuppressWarnings("NullableProblems")
+	@Override
 	public <T> List<Future<T>> invokeAll(Collection<? extends Callable<T>> tasks) throws InterruptedException {
 		return e.invokeAll(tasks);
 	}
 
-	@SuppressWarnings("NullableProblems")
+	@Override
 	public <T> List<Future<T>> invokeAll(Collection<? extends Callable<T>> tasks, long timeout, TimeUnit unit)
 			throws InterruptedException {
 		return e.invokeAll(tasks, timeout, unit);
 	}
 
-	@SuppressWarnings("NullableProblems")
+	@Override
 	public <T> T invokeAny(Collection<? extends Callable<T>> tasks)
 			throws InterruptedException, ExecutionException {
 		return e.invokeAny(tasks);
 	}
 
+	@Override
 	public <T> T invokeAny(Collection<? extends Callable<T>> tasks, long timeout, TimeUnit unit)
 			throws InterruptedException, ExecutionException, TimeoutException {
 		return e.invokeAny(tasks, timeout, unit);