+++ hutool-core/src/test/java/cn/hutool/core/date/ChineseDateTest.java
@@ -45,4 +45,27 @@ public class ChineseDateTest {
 		date = new ChineseDate(DateUtil.parseDate("1996-07-15"));
 		Assert.assertEquals("丙子鼠年 五月三十", date.toString());
 	}
+	@Test
+	public void getCyclicalYMDTest(){
+		//通过公历构建
+		ChineseDate chineseDate = new ChineseDate(DateUtil.parseDate("1993-01-06"));
+		String cyclicalYMD = chineseDate.getCyclicalYMD();
+		Assert.assertEquals("壬申年癸丑月丁亥日",cyclicalYMD);
+	}
+
+	@Test
+	public void getCyclicalYMDTest2(){
+		//通过农历构建
+		ChineseDate chineseDate = new ChineseDate(1992,12,14);
+		String cyclicalYMD = chineseDate.getCyclicalYMD();
+		Assert.assertEquals("壬申年癸丑月丁亥日",cyclicalYMD);
+	}
+
+	@Test
+	public void getCyclicalYMDTest3(){
+		//通过公历构建
+		ChineseDate chineseDate = new ChineseDate(DateUtil.parseDate("2020-08-28"));
+		String cyclicalYMD = chineseDate.getCyclicalYMD();
+		Assert.assertEquals("庚子年甲申月癸卯日",cyclicalYMD);
+	}
 }