+++ hutool-core/src/test/java/cn/hutool/core/util/TypeUtilTest.java
@@ -56,4 +56,5 @@ public class TypeUtilTest {
 		public void service(String string) {
 		}
 	}
+
 }