+++ hutool-crypto/src/test/java/cn/hutool/crypto/test/digest/HmacTest.java
@@ -1,4 +1,4 @@
-package cn.hutool.crypto.test;
+package cn.hutool.crypto.test.digest;
 
 import org.junit.Assert;
 import org.junit.Test;